+++ Look-ID/src/App.jsx
@@ -2,7 +2,7 @@ import NavBar from './components/navBar/NavBar'
 import './App.css'
 import Section from './components/section/Section'
 import Header from './components/header/Header'
-import AlunosCadastrados from './pages/AlunosCadastrados'
+import AlunosCadastrados from './components/pages/AlunosCadastrados'
 function App() {
   return (
     <main>
