+++ Look-ID/src/components/pages/AlunosComponent.jsx
@@ -1,4 +1,4 @@
-import Style from "../components/Styles/AlunosCadastrados.module.css";
+import Style from "../Styles/AlunosCadastrados.module.css";
 function AlunosComponent({nome,ra,turma}){
     return(
         <div className={Style.alunos}>
